Add AJAX on a arrayToJson, read-only for users

+++ resources/views/add.blade.php
@@ -58,18 +58,30 @@ echo '<form method="post" action="/user_upload">';
     unset($arrCol[0]);
     echo '<tr>' . PHP_EOL;
     echo '<td>' . $dep . '</td>';
-    foreach ($arrCol as $key => $colnum) {
-        if ($colnum == 1 && isset($arrAddRow[$key])) {
-            echo '<td><input type="text" pattern="' . $pattern . '" name="' . $arrAddRow[$key] . '" class="regex"></td>';
-        } elseif ($colnum > 1 && isset($arrAddRow[$key])) {
-            echo '<td colspan="' . $colnum . '"><input type="text" pattern="' . $pattern . '" name="' . $arrAddRow[$key] . '" class="regex"></td>';
+    if ($read_only == 'disabled') {
+        foreach ($arrCol as $key => $colnum) {
+            if ($colnum == 1 && isset($arrAddRow[$key])) {
+                echo '<td><input type="text" pattern="' . $pattern . '" name="' . $arrAddRow[$key] . '" class="regex"></td>';
+            } elseif ($colnum > 1 && isset($arrAddRow[$key])) {
+                echo '<td colspan="' . $colnum . '"><input type="text" pattern="' . $pattern . '" name="' . $arrAddRow[$key] . '" class="regex"></td>';
+            }
+        }
+    } else {
+        foreach ($arrCol as $key => $colnum) {
+            if ($colnum == 1 && isset($arrAddRow[$key])) {
+                echo '<td>' . $arrAddRow[$key] . '</td>';
+            } elseif ($colnum > 1 && isset($arrAddRow[$key])) {
+                echo '<td colspan="' . $colnum . '">' . $pattern . '</td>';
+            }
         }
     }
     $table_info = $name . ' + ' . $table_uuid . ' + ' . $row_uuid . ' + ' . $user_id . ' + ' . $dep;
     echo '<input type="hidden" name="table_information" value="' . $table_info . '"';
     echo '</tr>' . PHP_EOL;
     echo '</table>' . PHP_EOL;
-    echo '<input class="btn-submit-ae" type="button" value="Отправить" onclick="this.parentNode.submit();">';
+    if ($read_only == 'disabled') {
+       echo '<input class="btn-submit-ae" type="button" value="Отправить" onclick="this.parentNode.submit();">';
+    }
     echo '</form>' . PHP_EOL;
 @endphp
